Refactoring

+++ src/main/java/app/workout/WorkoutDto.java
@@ -6,13 +6,13 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Builder
-public record WorkoutDto(
+public record WorkoutDto (
          Long id,
          String activityId,
          String workoutType,
          String description,
          BigDecimal kcalBurned,
-         Double durationInMinutes,
+         BigDecimal durationInMinutes,
          int intensityLevel,
          LocalDate date
  ) {}
